refactor(ai-search-launcher): move FOCUS_VISIBLE_CLASSES and two icons to src/common
