+++ src/common/constants.ts
@@ -1,2 +1,5 @@
-export const QUERY_MAX_CHARACTER_LENGTH = 500;
+export const FOCUS_VISIBLE_CLASSES =
+  'focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-blue-600 dark:focus-visible:outline-blue-300';
+
 export const LEGACY_ENDPOINT = 'https://multimodal.search.rezolve.com';
+export const QUERY_MAX_CHARACTER_LENGTH = 500;
