+++ src/official-widgets/ai-search-launcher/components/FullScreenContainer.tsx
@@ -7,8 +7,8 @@ import ShadowWrapper from '../../../common/components/shadow-wrapper';
 import CloseIcon from '../../../common/icons/CloseIcon';
 import PlusCircleIcon from '../../../common/icons/PlusCircleIcon';
 import { WidgetBreakpoint } from '../../../common/types/constants';
-import { FOCUS_VISIBLE_CLASSES } from '../constants';
-import SpeakerIcon from '../icons/SpeakerIcon';
+import { FOCUS_VISIBLE_CLASSES } from '../../../common/constants';
+import SpeakerIcon from '../../../common/icons/SpeakerIcon';
 
 interface FullScreenContainerProps {
   open: boolean;
